man: sort formatter tokens by en_US.UTF-8 locale
LC_COLLATE=C would be an option as well, but I think this reads better.
There's not really any good/useful sort option here since the tokens are
arbitrary. I suspect most people pour over this list by way of text
search so really this only serves as a reverse mapping where the
explanation of the formatter token is more important for
discoverability.